% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrCook.R
\name{coef.dynrCook}
\alias{coef.dynrCook}
\title{Extract fitted parameters from a dynrCook Object}
\usage{
\method{coef}{dynrCook}(object, ...)
}
\arguments{
\item{object}{The dynrCook object for which the coefficients are desired}

\item{...}{further named arguments, ignored for this method}
}
\value{
A numeric vector of the fitted parameters.
}
\description{
Extract fitted parameters from a dynrCook Object
}
\examples{
# Let cookedModel be the output from dynr.cook
#coef(cookedModel)
}
\seealso{
Other S3 methods \code{\link{logLik.dynrCook}}
}

