% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrCook.R
\name{confint.dynrCook}
\alias{confint.dynrCook}
\title{Confidence Intervals for Model Parameters}
\usage{
\method{confint}{dynrCook}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{parm}{which parameters are to be given confidence intervals}

\item{level}{the confidence level}

\item{...}{further names arguments. Ignored.}
}
\value{
A matrix with columns giving lower and upper confidence limits for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 as a percentage (e.g. by default 2.5% and 97.5%).
}
\description{
Confidence Intervals for Model Parameters
}
\details{
The \code{parm} argument can be a numeric vector or a vector of names. If it is missing then it defaults to using all the parameters.

These are Wald-type confidence intervals based on the standard errors of the (transformed) parameters.  Wald-type confidence intervals are known to be inaccurate for variance parameters, particularly when the variance is near zero (See references for issues with Wald-type confidence intervals).
}
\examples{
# Let cookedModel be the output from dynr.cook
#confint(cookedModel)
}
\references{
Pritikin, J.N., Rappaport, L.M. & Neale, M.C.  (In Press). Likelihood-Based Confidence Intervals for a Parameter With an Upper or Lower Bound.  Structural Equation Modeling.  DOI: 10.1080/10705511.2016.1275969

Neale, M. C. & Miller M. B. (1997). The use of likelihood based confidence intervals in genetic models. Behavior Genetics, 27(2), 113-120.

Pek, J. & Wu, H. (2015). Profile likelihood-based confidence intervals and regions for structural equation models. Psychometrica, 80(4), 1123-1145.

Wu, H. & Neale, M. C. (2012). Adjusted confidence intervals for a bounded parameter. Behavior genetics, 42(6), 886-898.
}

