% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynR-package.R
\docType{package}
\name{dynR-package}
\alias{dynR}
\alias{dynR-package}
\title{dynR: Dynamic Connectivity Analysis for Neurophysiological Timeseries}
\description{
An R port of the Python dynfc library for computing dynamic functional
connectivity representations from multivariate neurophysiological
timeseries, including BOLD fMRI, EEG, LFP, and related signals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dynr.circadia-lab.uk}
  \item \url{https://github.com/circadia-bio/dynR}
  \item Report bugs at \url{https://github.com/circadia-bio/dynR/issues}
}

}
\author{
\strong{Maintainer}: Lucas França \email{lucas.franca@northumbria.ac.uk} (\href{https://orcid.org/0000-0003-0853-1319}{ORCID})

Authors:
\itemize{
  \item Lucas França \email{lucas.franca@northumbria.ac.uk} (\href{https://orcid.org/0000-0003-0853-1319}{ORCID})
  \item Mario Leocadio-Miguel \email{mario.miguel@northumbria.ac.uk} (\href{https://orcid.org/0000-0002-7248-3529}{ORCID})
  \item Dafnis Batallé \email{dafnis.batalle@kcl.ac.uk} (\href{https://orcid.org/0000-0003-2097-979X}{ORCID})
}

}
\keyword{internal}
