% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrFuncAddress.R
\name{dynr.config}
\alias{dynr.config}
\title{Check that dynr in configured properly}
\usage{
dynr.config(verbose = FALSE)
}
\arguments{
\item{verbose}{logical.  Whether to print messages during/after checks}
}
\value{
No return value.
}
\description{
Check that dynr in configured properly
}
\details{
The 'dynr' package requires additional set-up and configuration beyond
just installing the package.  In particular, it requires compiling C code
along with GSL to run (cook) models.  This function runs some basic checks
of the configuration.  We check that (1) R is on the PATH variable, (2)
Rtools exists and is on the PATH variable for Windows, (3) a C compiler
is available, and (4) GSL is available and on the PATH.

In general, see the 'Installation for Users' vignette for set-up and
configuration instructions.
}
\examples{
\dontrun{dynr.config()}
}

