% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrMi.R
\name{dynr.mi}
\alias{dynr.mi}
\title{Multiple Imputation of dynrModel objects}
\usage{
dynr.mi(model, m = 5, aux.variable, imp.obs = FALSE, imp.exo = FALSE, lag)
}
\arguments{
\item{model}{dynrModel object}

\item{m}{number of multiple imputations}

\item{aux.variable}{names of auxiliary variables used in imputation}

\item{imp.obs}{logical. whether to impute the observed variables}

\item{imp.exo}{logical. whether to impute the exogenous variables}

\item{lag}{numeric. the number of lags to use}
}
\description{
Multiple Imputation of dynrModel objects
}
\details{
This function is in alpha-testing form.  Please do not use or rely on it for now. A full implementation is in progress.
}

