% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrPlot.R
\name{dynr.plotFreq}
\alias{dynr.plotFreq}
\title{Plot of the estimated frequencies of the regimes across all individuals and time points
based on their smoothed regime probabilities}
\usage{
dynr.plotFreq(res, dynrModel, names.regime, title, xlab, ylab, textsize = 12,
  print = TRUE)
}
\arguments{
\item{res}{The dynr object returned by dynr.cook().}

\item{dynrModel}{The model object to plot.}

\item{names.regime}{(optional) Names of the regimes (must match the length of the number of regimes)}

\item{title}{(optional) Title of the plot.}

\item{xlab}{(optional) Label of the x-axis.}

\item{ylab}{(optional) Label of the y-axis.}

\item{textsize}{(default = 12) Text size for the axis labels and title (= textsize + 2).}

\item{print}{(default = TRUE) A flag for whether the plot should be printed.}
}
\description{
Plot of the estimated frequencies of the regimes across all individuals and time points
based on their smoothed regime probabilities
}

