% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrCook.R
\docType{class}
\name{dynrCook-class}
\alias{$,dynrCook-method}
\alias{dynrCook-class}
\alias{dynrDebug-class}
\alias{print,dynrCook-method}
\alias{show,dynrCook-method}
\title{The dynrCook Class}
\description{
The dynrCook Class
}
\details{
This is an internal class structure.  You should not use it directly.
Use \code{\link{dynr.cook}} instead.
}

