% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrRecipe.R
\docType{class}
\name{dynrDynamics-class}
\alias{dynrDynamics-class}
\alias{dynrDynamicsFormula-class}
\alias{dynrDynamicsMatrix-class}
\title{The dynrDynamics Class}
\description{
The dynrDynamics Class
}
\details{
This is an internal class structure.  The classes \code{dynrDynamicsFormula-class} 
and \code{dynrDynamicsMatrix-class} are subclasses of this.  However, you should 
not use it directly.
Use \code{\link{prep.matrixDynamics}} or \code{\link{prep.formulaDynamics}} instead.
}

