% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrRecipe.R
\docType{class}
\name{dynrInitial-class}
\alias{dynrInitial-class}
\title{The dynrInitial Class}
\description{
The dynrInitial Class
}
\details{
This is an internal class structure.  You should not use it directly.
Use \code{\link{prep.initial}} instead.
}

