% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrRecipe.R
\docType{class}
\name{dynrMeasurement-class}
\alias{dynrMeasurement-class}
\title{The dynrMeasurement Class}
\description{
The dynrMeasurement Class
}
\details{
This is an internal class structure.  You should not use it directly.
Use \code{\link{prep.measurement}} or \code{\link{prep.loadings}} instead.
}

