% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrRecipe.R
\docType{class}
\name{dynrNoise-class}
\alias{dynrNoise-class}
\title{The dynrNoise Class}
\description{
The dynrNoise Class
}
\details{
This is an internal class structure.  You should not use it directly.
Use \code{\link{prep.noise}} instead.
}

