% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrRecipe.R
\docType{class}
\name{dynrRecipe-class}
\alias{$,dynrRecipe-method}
\alias{dynrRecipe-class}
\title{The dynrRecipe Class}
\description{
The dynrRecipe Class
}
\details{
This is an internal class structure.  You should not use it directly.
The following are all subclasses of this class: \code{\link{dynrMeasurement-class}},
\code{\link{dynrDynamics-class}}, \code{\link{dynrRegimes-class}},
\code{\link{dynrInitial-class}}, \code{\link{dynrNoise-class}},
and \code{\link{dynrTrans-class}}.  Recipes are the things that
go into a \code{\link{dynrModel-class}} using \code{\link{dynr.model}}.
Use the recipe prep functions (\code{\link{prep.measurement}},
\code{\link{prep.formulaDynamics}}, \code{\link{prep.matrixDynamics}},
\code{\link{prep.regimes}}, \code{\link{prep.initial}}, \code{\link{prep.noise}},
or \code{\link{prep.tfun}}) to create these classes instead.
}

