% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrRecipe.R
\docType{class}
\name{dynrRegimes-class}
\alias{dynrRegimes-class}
\title{The dynrRegimes Class}
\description{
The dynrRegimes Class
}
\details{
This is an internal class structure.  You should not use it directly.
Use \code{\link{prep.regimes}} instead.
}

