% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrRecipe.R
\docType{class}
\name{dynrTrans-class}
\alias{dynrTrans-class}
\title{The dynrTrans Class}
\description{
The dynrTrans Class
}
\details{
This is an internal class structure.  You should not use it directly.
Use \code{\link{prep.tfun}} instead.
}

