% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrCook.R
\name{logLik.dynrCook}
\alias{logLik.dynrCook}
\title{Extract the log likelihood from a dynrCook Object}
\usage{
\method{logLik}{dynrCook}(object, ...)
}
\arguments{
\item{object}{The dynrCook object for which the log likelihood is desired}

\item{...}{further named arguments, ignored for this method}
}
\value{
An object of class \code{logLik}.
}
\description{
Extract the log likelihood from a dynrCook Object
}
\details{
The 'df' attribute for this object is the number of freely estimated parameters. The 'nobs' attribute is the total number of rows of data, adding up the number of time points for each person.
}
\examples{
# Let cookedModel be the output from dynr.cook
#logLik(cookedModel)
}
\seealso{
Other S3 methods \code{\link{coef.dynrCook}}
}

