% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrCook.R
\docType{methods}
\name{names,dynrCook-method}
\alias{names,dynrCook-method}
\title{Extract the free parameter names of a dynrCook object}
\usage{
\S4method{names}{dynrCook}(x)
}
\arguments{
\item{x}{The dynrCook object from which the free parameter names are desired}
}
\description{
Extract the free parameter names of a dynrCook object
}

