% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrModel.R
\docType{methods}
\name{names,dynrModel-method}
\alias{names,dynrModel-method}
\title{Extract the free parameter names of a dynrModel object}
\usage{
\S4method{names}{dynrModel}(x)
}
\arguments{
\item{x}{The dynrModel object from which the free parameter names are desired}
}
\description{
Extract the free parameter names of a dynrModel object
}

