% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrModel.R
\name{nobs.dynrModel}
\alias{nobs.dynrModel}
\title{Extract the number of observations for a dynrModel object}
\usage{
\method{nobs}{dynrModel}(object, ...)
}
\arguments{
\item{object}{An unfitted model object}

\item{...}{Further named arguments. Ignored.}
}
\value{
A single number. The total number of observations across all IDs.
}
\description{
Extract the number of observations for a dynrModel object
}
\details{
We return the total number of rows of data, adding up the number of time points for each person. For some purposes, you may want the mean number of observations per person or the number of people instead.  These are not currently supported via \code{nobs}.
}
\examples{
# Let rawModel be the output from dynr.model
#nobs(rawModel)
}

