% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{oscData}
\alias{oscData}
\title{Another simulated multilevel multi-subject time series of a damped oscillator model}
\format{A data frame with 1,800 rows and 6 variables}
\usage{
data(oscData)
}
\description{
The variables are as follows:
}
\details{
\itemize{
  \item id. Person ID
  \item times. Continuous time of measurement
  \item y1. Observed score 1
  \item u1. Covariate 1
  \item u2. Covariate 2
  \item trueb. True value of person-specific random effect	
}
}
\keyword{datasets}

