% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{Oscillator}
\alias{Oscillator}
\title{Simulated time series data of a damped linear oscillator}
\format{A data frame with 1000 rows and 5 variables}
\usage{
data(Oscillator)
}
\description{
A dataset simulated using a damped linear oscillator model in continuous time
with 1 observed indicator for identifying two latent factors (position and velocity).
The variables are as follows:
}
\details{
\itemize{
  \item id. ID of the systems (1 to 1 because this is a single person)
  \item y1. Noisy observed position
  \item times. Time index (1000 time points) spaced at one unit intervals
  \item x1. True latent position
  \item x2. True latent velocity
}
}
\keyword{datasets}

