% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrPlot.R
\name{plotFormula}
\alias{plotFormula}
\title{Plot the formula from a model}
\usage{
plotFormula(dynrModel, ParameterAs, printDyn = TRUE, printMeas = TRUE,
  textsize = 4)
}
\arguments{
\item{dynrModel}{The model object to plot.}

\item{ParameterAs}{The parameter values or names to plot. The underscores in parameter names are saved for
use of subscripts.  Greek letters can be specified as corresponding LaTeX symbols without backslashes (e.g., "lambda")
and printed as greek letters.}

\item{printDyn}{A logical value indicating whether or not to plot the formulas for the dynamic model.}

\item{printMeas}{A logical value indicating whether or not to plot the formulas for the measurement model}

\item{textsize}{The text size use in the plot.}
}
\description{
Plot the formula from a model
}
\details{
This function typesets a set of formulas that represent the model.
}

