% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{PPsim}
\alias{PPsim}
\title{Simulated time series data for multiple eco-systems based on a predator-and-prey model}
\format{A data frame with 1000 rows and 6 variables}
\usage{
data(PPsim)
}
\description{
A dataset simulated using a continuous-time nonlinear predator-and-prey model
with 2 observed indicators for identifying two latent factors. The variables are as follows:
}
\details{
\itemize{
  \item id. ID of the systems (1 to 20)
  \item time. Time index (50 time points for each system)
  \item prey. The true population of the prey species
  \item predator. The true population of the predator species
  \item x. Observed indicator for the population of the prey species
  \item y. Observed indicator for the population of the predator species
}
}
\keyword{datasets}

