% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrRecipe.R
\name{printex}
\alias{printex}
\alias{printex,dynrCook-method}
\alias{printex,dynrDynamicsFormula-method}
\alias{printex,dynrDynamicsMatrix-method}
\alias{printex,dynrInitial-method}
\alias{printex,dynrMeasurement-method}
\alias{printex,dynrModel-method}
\alias{printex,dynrNoise-method}
\alias{printex,dynrRegimes-method}
\title{The printex Method}
\usage{
printex(object, show = TRUE, AsMatrix = TRUE, ...)
}
\arguments{
\item{object}{The dynr object (recipe, model, cooked, model)}

\item{show}{logical. Whether or not to show the result in the console.}

\item{AsMatrix}{logical. Whether to put the object in matrix form.}

\item{...}{Further named arguments, passed to internal method.}
}
\description{
The printex Method
}
\details{
This is a general way of getting a LaTeX string for recipes,
models, and cooked models.  It is a great way to check that
you specified the model or recipe you think you did before
estimating its free parameters (cooking).  After the model
is cooked, you can use it to get LaTeX code with the estimated
parameters in it.  The underscores in parameter names are saved
for use of subscripts.  Greek letters can be specified as
corresponding LaTeX symbols without backslashes (e.g., "lambda")
and printed as greek letters.

Further arguments for the \code{dynrModel} method are with their
defaults are \code{ParameterAs}, \code{printDyn=TRUE},
\code{printMeas=TRUE}, \code{printInit=FALSE}, \code{printRS=FALSE},
and \code{outFile}.
}

