\name{rscdekf-package}
\alias{rscdekf-package}
\alias{rscdekf}
\docType{package}
\title{
\packageTitle{rscdekf}
}
\description{
\packageDescription{rscdekf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rscdekf}
\packageIndices{rscdekf}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rscdekf}

Maintainer: \packageMaintainer{rscdekf}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
