% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{RSPPsim}
\alias{RSPPsim}
\title{Simulated time series data for multiple eco-systems based on a regime-switching predator-and-prey model}
\format{A data frame with 6000 rows and 8 variables}
\usage{
data(RSPPsim)
}
\description{
A dataset simulated using a regime-switching continuous-time nonlinear predator-and-prey model
with 2 observed indicators for identifying two latent factors. The variables are as follows:
}
\details{
\itemize{
  \item id. ID of the systems (1 to 20)
  \item time. Time index (300 time points for each system)
  \item prey. The true population of the prey species
  \item predator. The true population of the predator species
  \item x. Observed indicator for the population of the prey species
  \item y. Observed indicator for the population of the predator species
  \item cond. A time-varying covariate indicating the conditions of the respective eco-system across time which 
  affects the regime-switching transition matrix
  \item regime. The true regime indicators across time (1 and 2).
}
}
\keyword{datasets}

