% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrRecipe.R
\name{substituteFormula}
\alias{substituteFormula}
\title{A internal-use only function for substituting formula. If the RHS of \code{formula} has terms in the LHS of \code{term.formula}, this function replaces any appearance with the RHS of \code{term.formula}}
\usage{
substituteFormula(formula, term.formula)
}
\arguments{
\item{formula}{a list of original formulas}

\item{term.formula}{a list of term formulas}
}
\value{
a list of formulas after the replacement
}
\description{
A internal-use only function for substituting formula. If the RHS of \code{formula} has terms in the LHS of \code{term.formula}, this function replaces any appearance with the RHS of \code{term.formula}
}
\examples{
#substitutedformula <- substituteFormula(formula, term.formula)
}

