% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrCook.R
\docType{methods}
\name{summary,dynrCook-method}
\alias{summary,dynrCook-method}
\title{Get the summary of a dynrCook object}
\usage{
\S4method{summary}{dynrCook}(object)
}
\arguments{
\item{object}{The dynrCook object for which the summary is desired.}
}
\description{
Get the summary of a dynrCook object
}
\details{
The summary gives information on the free parameters estimated: names, parameter values, numerical Hessian-based standard errors, t-values (values divided by standard errors), and standard-error based confidence intervals.  Additionally, the likelihood, AIC, and BIC are provided.
}

