% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrCook.R
\name{summary.dynrCook}
\alias{summary.dynrCook}
\title{Get the summary of a dynrCook object}
\usage{
\method{summary}{dynrCook}(object, ...)
}
\arguments{
\item{object}{The dynrCook object for which the summary is desired.}

\item{...}{Further named arguments, passed to the print method (e.g., \code{digits} and \code{signif.stars}).}
}
\value{
Object of class summary.dynrCook.  Primarily used for showing the results of a fitted model.
}
\description{
Get the summary of a dynrCook object
}
\details{
The summary gives information on the free parameters estimated: names, parameter values, numerical Hessian-based standard errors, t-values (values divided by standard errors), and standard-error based confidence intervals.  Additionally, the likelihood, AIC, and BIC are provided.

Note that an exclamation point (!) in the final column of the summary table indicates that
the standard error and confidence interval for this parameter may not be trustworthy. The corresponding
element of the (transformed, inverse) Hessian was negative and an absolute value was taken to make it positive.
}

