% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrCook.R
\name{vcov.dynrCook}
\alias{vcov.dynrCook}
\title{Extract the Variance-Covariance Matrix of a dynrCook object}
\usage{
\method{vcov}{dynrCook}(object, ...)
}
\arguments{
\item{object}{The dynrCook object for which the variance-covariance matrix is desired}

\item{...}{further named arguments, ignored by this method}
}
\description{
Extract the Variance-Covariance Matrix of a dynrCook object
}
\details{
This is the inverse Hessian of the transformed parameters.
}

