% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrModel.R, R/dynrCook.R
\name{coef.dynrModel}
\alias{coef.dynrCook}
\alias{coef.dynrModel}
\alias{coef<-}
\alias{coef<-.dynrModel}
\title{Extract fitted parameters from a dynrCook Object}
\usage{
\method{coef}{dynrModel}(object, ...)

coef(object) <- value

\method{coef}{dynrModel}(object) <- value

\method{coef}{dynrCook}(object, ...)
}
\arguments{
\item{object}{The dynrCook object for which the coefficients are desired}

\item{...}{further named arguments, ignored for this method}

\item{value}{values for setting}
}
\value{
A numeric vector of the fitted parameters.
}
\description{
aliases coef.dynrModel coef<- coef<-.dynrModel
}
\seealso{
Other S3 methods \code{\link{logLik.dynrCook}}
}

