\name{dynr-package}
\alias{dynr-package}
\alias{dynr}
\docType{package}
\title{
\packageTitle{dynr}
}
\description{
\packageDescription{dynr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dynr}
\packageIndices{dynr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{dynr}

Maintainer: \packageMaintainer{dynr}
}
\references{
~~ Literature or other references for background information ~~
}
\note{
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ Other packages ~~ %\code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
#~~ simple examples of the most important functions ~~
}
