% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrCook.R
\name{dynr.cook}
\alias{dynr.cook}
\title{Cook a dynr model to estimate its free parameters}
\usage{
dynr.cook(dynrModel, conf.level = 0.95, infile, verbose = TRUE,
  debug_flag = FALSE)
}
\arguments{
\item{dynrModel}{a dynr model compiled using dynr.model, consisting of recipes for submodels, starting values, parameter names, and C code for each submodel}

\item{conf.level}{a cumulative proportion indicating the level of desired confidence intervals for the final parameter estimates (default is .95)}

\item{infile}{(not required for models specified through the recipe functions) the name of a file that has the C codes for all dynr submodels for those interested in specifying a model directly in C}

\item{verbose}{a flag (TRUE/FALSE) indicating whether more detailed intermediate output during the estimation process should be printed}

\item{debug_flag}{a flag (TRUE/FALSE) indicating whether users want additional dynr output that can be used for diagnostic purposes}
}
\description{
Cook a dynr model to estimate its free parameters
}
\details{
TO BE COMPLETED:
for "cooking dinner" -- namely, to start the estimation process
a description of things output when debug_flag = FALSE
a description of things output when debug_flag = TRUE
}
\examples{
#fitted.model <- dynr.cook(model)
}
\seealso{
\code{\link{dynr.cook}}
}

