% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrPlot.R
\name{dynr.ggplot}
\alias{dynr.ggplot}
\title{The ggplot of the smoothed state estimates and the most likely regimes}
\usage{
dynr.ggplot(x, data.dynr, states, names.state = paste0("state", states),
  title = "Smoothed State Values", numSubjDemo = 2)
}
\arguments{
\item{x}{The dynr object returned by dynr.run().}

\item{data.dynr}{The dynr data returned by dynr.data().}

\item{states}{The indices of the states to be plotted.}

\item{names.state}{The names of the states to be plotted.}

\item{title}{A title of the plot.}

\item{numSubjDemo}{The number of subjects to be randomly selected for plotting.}
}
\description{
The ggplot of the smoothed state estimates and the most likely regimes
}

