% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrPlot.R
\name{dynr.ggplot}
\alias{dynr.ggplot}
\title{The ggplot of the smoothed state estimates and the most likely regimes}
\usage{
dynr.ggplot(res, data.dynr, numSubjDemo = 2, states, names.state,
  names.regime, shape.values, title = "Smoothed State Values",
  ylab = "Smoothed State Values", is.bw = FALSE, colorPalette = "Set2",
  fillPalette = "Set2", mancolorPalette, manfillPalette, ...)
}
\arguments{
\item{res}{The dynr object returned by dynr.run().}

\item{data.dynr}{The dynr data returned by dynr.data().}

\item{numSubjDemo}{The number of subjects to be randomly selected for plotting.}

\item{states}{The indices of the states to be plotted.}

\item{names.state}{The names of the states to be plotted, which can be missing.}

\item{names.regime}{The names of the regimes to be plotted, which can be missing.}

\item{shape.values}{A vector of values that correspond to the shapes of the points, which can be missing.}

\item{title}{A title of the plot.}

\item{ylab}{The label of the y axis.}

\item{is.bw}{Is plot in black and white?}

\item{colorPalette}{A palette function for lines and dots that when called with a single integer argument (the number of levels in the scale) returns the values that they should take.}

\item{fillPalette}{A palette function for blocks that when called with a single integer argument (the number of levels in the scale) returns the values that they should take.}

\item{mancolorPalette}{A color palette for manually scaling the colors of lines and plots.}

\item{manfillPalette}{A color palette for manually scaling the colors of filled blocks.}

\item{...}{A list of element name, element pairings that modify the existing ggplot theme. Consult the theme() function in the R package ggplot.}
}
\description{
The ggplot of the smoothed state estimates and the most likely regimes
}

