% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrModel.R
\name{dynr.model}
\alias{dynr.model}
\title{Create the model for dynr}
\usage{
dynr.model(num_regime = 1, dim_latent_var, xstart, ub, lb,
  options = default.model.options)
}
\arguments{
\item{num_regime}{An integer number of the regimes.}

\item{dim_latent_var}{An integer number of the latent variables.}

\item{xstart}{The starting values for parameter estimation.}

\item{ub}{The upper bounds of the estimated parameters.}

\item{lb}{The lower bounds of the estimated parameters.}

\item{options}{A list of NLopt estimation options. By default, xtol_rel=1e-7, stopval=-9999, ftol_rel=-1, ftol_abs=-1, maxeval=as.integer(-1), and maxtime=-1.}
}
\value{
A list of model statements to be passed to dynr.run().
}
\description{
Create the model for dynr
}

