% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynrCook.R
\docType{class}
\name{dynrCook-class}
\alias{dynrCook-class}
\alias{dynrDebug-class}
\alias{dynrOutall-class}
\title{The dynrCook Class}
\description{
The dynrCook Class
}
\details{
This is an internal class structure.  You should not use it directly.
Use \code{\link{dynr.cook}} instead.
}

