% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrModel.R
\docType{class}
\name{dynrModel-class}
\alias{$,dynrModel-method}
\alias{$<-,dynrModel-method}
\alias{dynrModel-class}
\alias{print,dynrModel-method}
\alias{show,dynrModel-method}
\title{The dynrModel Class}
\description{
The dynrModel Class
}
\details{
This is an internal class structure.  You should not use it directly.
Use \code{\link{dynr.model}} instead.
}

