% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrCook.R
\name{logLik.dynrCook}
\alias{deviance.dynrCook}
\alias{logLik.dynrCook}
\title{Extract the log likelihood from a dynrCook Object}
\usage{
\method{logLik}{dynrCook}(object, ...)

\method{deviance}{dynrCook}(object, ...)
}
\arguments{
\item{object}{The dynrCook object for which the log likelihood is desired}

\item{...}{further named arguments, ignored for this method}
}
\value{
In the case of \code{logLik}, an object of class \code{logLik}.
}
\description{
Extract the log likelihood from a dynrCook Object
}
\details{
The 'df' attribute for this object is the number of freely estimated parameters. The 'nobs' attribute is the total number of rows of data, adding up the number of time points for each person.

The \code{deviance} method returns minus two times the log likelihood.
}
\examples{
# Minimal model
require(dynr)

meas <- prep.measurement(
	values.load=matrix(c(1, 0), 1, 2),
	params.load=matrix(c('fixed', 'fixed'), 1, 2),
	state.names=c("Position","Velocity"),
	obs.names=c("y1"))

ecov <- prep.noise(
	values.latent=diag(c(0, 1), 2),
	params.latent=diag(c('fixed', 'dnoise'), 2),
	values.observed=diag(1.5, 1),
	params.observed=diag('mnoise', 1))

initial <- prep.initial(
	values.inistate=c(0, 1),
	params.inistate=c('inipos', 'fixed'),
	values.inicov=diag(1, 2),
	params.inicov=diag('fixed', 2))

dynamics <- prep.matrixDynamics(
	values.dyn=matrix(c(0, -0.1, 1, -0.2), 2, 2),
	params.dyn=matrix(c('fixed', 'spring', 'fixed', 'friction'), 2, 2),
	isContinuousTime=TRUE)

data(Oscillator)
data <- dynr.data(Oscillator, id="id", time="times", observed="y1")

model <- dynr.model(dynamics=dynamics, measurement=meas,
	noise=ecov, initial=initial, data=data)

cook <- dynr.cook(model,
	verbose=FALSE, optimization_flag=FALSE, hessian_flag=FALSE)

# Now get the log likelihood!
logLik(cook)
}
\seealso{
Other S3 methods \code{\link{coef.dynrCook}}
}

