% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrCook.R
\name{nobs.dynrCook}
\alias{nobs.dynrCook}
\title{Extract the number of observations for a dynrCook object}
\usage{
\method{nobs}{dynrCook}(object, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{...}{Further named arguments. Ignored.}
}
\value{
A single number. The total number of observations across all IDs.
}
\description{
Extract the number of observations for a dynrCook object
}
\details{
We return the total number of rows of data, adding up the number of time points for each person. For some purposes, you may want the mean number of observations per person or the number of people instead.  These are not currently supported via \code{nobs}.
}

